requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
djangorestframework==3.9.0
idna==2.8
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pymessenger==0.0.7.0
pytz==2019.1
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
